Added mine full check, fixed save issues

+++ src/main/java/tech/mcprison/prison/mines/Mine.java
@@ -79,7 +79,7 @@ public class Mine {
         if (hasSpawn) {
             spawn = new Location(worldOptional.get(), (double) document.get("spawnX"),
                 (double) document.get("spawnY"), (double) document.get("spawnZ"),
-                (float) document.get("spawnPitch"), (float) document.get("spawnYaw"));
+                ((Double) document.get("spawnPitch")).floatValue(), ((Double) document.get("spawnYaw")).floatValue());
         }
 
         worldName = worldOptional.get().getName();
